pyproject.toml
[tool.poetry]
authors = ["Fernando Flores <ffvillaca07@gmail.com>"]
description = ""
name = "capstone"
# readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
Django = "*"
Pillow = "^10.0.0"
django-environ = "^0.11.2"
django-filter = "^23.2"
django-guardian = "^2.4.0"
django-vite = {git = "https://github.com/FFV47/django-vite.git"}
djangorestframework = "*"
drf-standardized-errors = {extras = ["openapi"], version = "^0.12.5"}
markdown = "^3.4.3"
orjson = "*"
pygments = "^2.15.1"
python = "^3.11"
python-magic = "^0.4.27"
pyyaml = "^6.0"
uritemplate = "^4.1.1"
whitenoise = "*"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
django-browser-reload = "^1.6.0"
django-debug-toolbar = "^4.0.0"
django-extensions = "^3.2.1"
django-stubs = {extras = ["compatible-mypy"], version = "*"}
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "*"}
djlint = "^1.12.1"
drf-spectacular = "^0.26.4"
pyopenssl = "^23.0.0"
pytest-django = "^4.5.2"
pytest-xdist = "^3.3.1"

[tool.djlint]
blank_line_after_tag = "load,extends,include,endblock"
format_attribute_template_tags = true
format_js = true
ignore = "H031,D018,H030,H017,T002,H014"
indent = 2
max_attribute_length = 80
max_line_length = 100
profile = "django"
use_gitignore = true

[tool.djlint.js]
indent_size = 2

[tool.ruff]
ignore = ["E501"]
line-length = 100

[tool.black]
line-length = 100

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
